Client can handle results of different data types
To allow complete use of Voyage's API, the Embedding Response
object is made generic (limited to float32, int8, uint8).

Additionally, the Embedding object is made into a union type
to allow it to comprehend default encoding (array of values)
and base64 encoding, which returns a string value. For the
latter case we provide a method AsSlice() that returns the
normal encoding if present, else decodes the base64 string
and converts to a slice when that is what was returned by the
model.

There are a couple of minor breaking changes:

1. Since the response is made generic, existing variables
   that declare a type may need to be updated
2. The numeric slice now needs to be retrieved as `AsNumeric`
   from the embedding object.